+++ src/LiveComponent/tests/Functional/Controller/BatchActionControllerTest.php
@@ -27,7 +27,7 @@ final class BatchActionControllerTest extends KernelTestCase
 
     public function testCanBatchActions(): void
     {
-        $dehydrated = $this->dehydrateComponent($this->mountComponent('with_actions'));
+        $dehydrated = $this->dehydrateComponent($this->mountComponent('with_actions'))->all();
 
         $this->browser()
             ->throwExceptions()
@@ -35,9 +35,13 @@ final class BatchActionControllerTest extends KernelTestCase
             ->assertSuccessful()
             ->assertSee('initial')
             ->use(function (HtmlResponse $response, KernelBrowser $browser) {
+                $rootElement = $response->crawler()->filter('ul')->first();
+                $liveData = json_decode($rootElement->attr('data-live-data-value'), true);
+                $liveProps = json_decode($rootElement->attr('data-live-props-value'), true);
+
                 $browser->post('/_components/with_actions/add', [
                     'json' => [
-                        'data' => json_decode($response->crawler()->filter('ul')->first()->attr('data-live-data-value')),
+                        'data' => $liveData + $liveProps,
                         'args' => ['what' => 'first'],
                     ],
                     'headers' => ['X-CSRF-TOKEN' => $response->crawler()->filter('ul')->first()->attr('data-live-csrf-value')],
@@ -46,9 +50,13 @@ final class BatchActionControllerTest extends KernelTestCase
             ->assertSee('initial')
             ->assertSee('first')
             ->use(function (HtmlResponse $response, KernelBrowser $browser) {
+                $rootElement = $response->crawler()->filter('ul')->first();
+                $liveData = json_decode($rootElement->attr('data-live-data-value'), true);
+                $liveProps = json_decode($rootElement->attr('data-live-props-value'), true);
+
                 $browser->post('/_components/with_actions/_batch', [
                     'json' => [
-                        'data' => json_decode($response->crawler()->filter('ul')->first()->attr('data-live-data-value')),
+                        'data' => $liveData + $liveProps,
                         'actions' => [
                             ['name' => 'add', 'args' => ['what' => 'second']],
                             ['name' => 'add', 'args' => ['what' => 'third']],
@@ -68,7 +76,7 @@ final class BatchActionControllerTest extends KernelTestCase
 
     public function testCsrfTokenIsChecked(): void
     {
-        $dehydrated = $this->dehydrateComponent($this->mountComponent('with_actions'));
+        $dehydrated = $this->dehydrateComponent($this->mountComponent('with_actions'))->all();
 
         $this->browser()
             ->post('/_components/with_actions/_batch', ['json' => [
@@ -81,7 +89,7 @@ final class BatchActionControllerTest extends KernelTestCase
 
     public function testRedirect(): void
     {
-        $dehydrated = $this->dehydrateComponent($this->mountComponent('with_actions'));
+        $dehydrated = $this->dehydrateComponent($this->mountComponent('with_actions'))->all();
 
         $this->browser()
             ->throwExceptions()
@@ -89,9 +97,13 @@ final class BatchActionControllerTest extends KernelTestCase
             ->assertSuccessful()
             ->interceptRedirects()
             ->use(function (HtmlResponse $response, KernelBrowser $browser) {
+                $rootElement = $response->crawler()->filter('ul')->first();
+                $liveData = json_decode($rootElement->attr('data-live-data-value'), true);
+                $liveProps = json_decode($rootElement->attr('data-live-props-value'), true);
+
                 $browser->post('/_components/with_actions/_batch', [
                     'json' => [
-                        'data' => json_decode($response->crawler()->filter('ul')->first()->attr('data-live-data-value')),
+                        'data' => $liveData + $liveProps,
                         'actions' => [
                             ['name' => 'add', 'args' => ['what' => 'second']],
                             ['name' => 'redirect'],
@@ -107,15 +119,19 @@ final class BatchActionControllerTest extends KernelTestCase
 
     public function testException(): void
     {
-        $dehydrated = $this->dehydrateComponent($this->mountComponent('with_actions'));
+        $dehydrated = $this->dehydrateComponent($this->mountComponent('with_actions'))->all();
 
         $this->browser()
             ->get('/_components/with_actions', ['json' => ['data' => $dehydrated]])
             ->assertSuccessful()
             ->use(function (HtmlResponse $response, KernelBrowser $browser) {
+                $rootElement = $response->crawler()->filter('ul')->first();
+                $liveData = json_decode($rootElement->attr('data-live-data-value'), true);
+                $liveProps = json_decode($rootElement->attr('data-live-props-value'), true);
+
                 $browser->post('/_components/with_actions/_batch', [
                     'json' => [
-                        'data' => json_decode($response->crawler()->filter('ul')->first()->attr('data-live-data-value')),
+                        'data' => $liveData + $liveProps,
                         'actions' => [
                             ['name' => 'add', 'args' => ['what' => 'second']],
                             ['name' => 'exception'],
@@ -132,15 +148,19 @@ final class BatchActionControllerTest extends KernelTestCase
 
     public function testCannotBatchWithNonLiveAction(): void
     {
-        $dehydrated = $this->dehydrateComponent($this->mountComponent('with_actions'));
+        $dehydrated = $this->dehydrateComponent($this->mountComponent('with_actions'))->all();
 
         $this->browser()
             ->get('/_components/with_actions', ['json' => ['data' => $dehydrated]])
             ->assertSuccessful()
             ->use(function (HtmlResponse $response, KernelBrowser $browser) {
+                $rootElement = $response->crawler()->filter('ul')->first();
+                $liveData = json_decode($rootElement->attr('data-live-data-value'), true);
+                $liveProps = json_decode($rootElement->attr('data-live-props-value'), true);
+
                 $browser->post('/_components/with_actions/_batch', [
                     'json' => [
-                        'data' => json_decode($response->crawler()->filter('ul')->first()->attr('data-live-data-value')),
+                        'data' => $liveData + $liveProps,
                         'actions' => [
                             ['name' => 'add', 'args' => ['what' => 'second']],
                             ['name' => 'nonLive'],